fixing component naming + moving MissionPlan to mission.py
