+++ GEMstack/knowledge/defaults/computation_graph.yaml
@@ -52,13 +52,10 @@ components:
   - driving_logic:
       inputs: 
       outputs: intent
-  - parking_component: # one way
+  - _mission_planner: # one way
       inputs: all
       outputs: mission_plan
-  - summon_component: # one way
-      inputs: all
-      outputs: route
-  - route_planning_component: # one way
+  - _route_planner: # one way
       inputs: all
       outputs: route
   - motion_planning:
