+++ GEMstack/state/mission.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
 from ..utils.serialization import register
 from enum import Enum
+from .route import PlannerEnum
+
 
 class MissionEnum(Enum):
     IDLE = 0            # not driving, no mission
@@ -14,4 +16,15 @@ class MissionEnum(Enum):
 @register
 class MissionObjective:
     type : MissionEnum = MissionEnum.IDLE
+    
+@dataclass
+@register
+class MissionPlan:
+    goal_x: float
+    goal_y: float
+    goal_orientation: float
+    planner_type : PlannerEnum = PlannerEnum.RRT_STAR
+    # other mission-specific parameters can be added here
+
+
