+++ launch/parking_detection.yaml
@@ -30,8 +30,8 @@ drive:
         parking_detection: parking_detection.ParkingSpotsDetector3D
         perception_normalization : StandardPerceptionNormalizer
     planning:
-        parking_component:
-            type: ParkingSim
+        _mission_planner:
+            type: parking_component.ParkingSim
 #usually can keep this constant
 computation_graph: !include "../GEMstack/knowledge/defaults/computation_graph.yaml"
 
@@ -75,10 +75,10 @@ variants:
                     # agent_detection : cone_detection_simple.ConeDetectorSimple3D
                     parking_detection: parking_detection.ParkingSpotsDetector3D
                 planning:
-                    parking_component:
-                        type: ParkingSim
-                    route_planning_component:
-                        type: RoutePlanningComponent
+                    _mission_planner:
+                        type: parking_component.ParkingSim
+                    _route_planner:
+                        type: route_planning_component.RoutePlanningComponent
                 
     sim:
         run:
